store sessions in dict, run delete in a separate thread
